Validate scheduling fields in basic ops scenario
Currently there is no validation of node scheduling fields - resource
class and traits - in the scenario tests. This change adds validation of
these fields to the bare metal basic ops test.

We query the flavor used to boot the instance, and extract all requested
resources and traits from extra_specs. These are matched against the
resource class and traits set on the bare metal node that was scheduled.

Change-Id: I9ddc895ead61cf02c6967ead094d061cb7f558d8
Related-Bug: #1722194